chore(lint): enable new PMD/Checkstyle rules + fix Sonar-only bugs
PMD additions (config/pmd/ruleset.xml): AvoidDuplicateLiterals (threshold 4),
ExcessiveParameterList (minimum 8), UselessParentheses. Checkstyle additions
(config/checkstyle/checkstyle.xml): TodoComment, HiddenField. Baseline
ceilings bumped in gradle.properties to absorb the new warning surface.

Sonar-only bug fixes that no local rule would have caught:
- BitmapData equals/hashCode/toString now use Arrays.* (records use
  identity on the int[] field, not content).
- RegionRleCodec.getBitFragment masks with 0xff to prevent sign extension
  when extractFrom is a negative byte.
- WeaponsLogic.proximityRadiusForLevel casts to double before addition.
- ConfigRegistry.slots → slotsByType (BLOCKER: clashed with SLOTS const).

Small cleanups: removed unused throws/parameters, replaced Math.random()
with ThreadLocalRandom, parameterised raw ContactSystem, dropped dangling
javadoc, and renamed snake_case AudioTypes methods to camelCase.

+++ infinity-server/src/main/java/infinity/systems/ship/WeaponsLogic.java
@@ -19,7 +19,7 @@ final class WeaponsLogic {
 
     /** Proximity radius: additive per level (L1=base, L2=base+1, …) — REFERENCE.md ## Bomb. */
     static double proximityRadiusForLevel(final int baseTiles, final int level) {
-        return baseTiles + (level - 1);
+        return (double) baseTiles + (level - 1);
     }
 
     /** Subspace velocity × scale, clamped to ±{@code maxJmeAbs}; sign preserved for backward-firing presets. */
